implemented interface with Qt GUI framework to perform data and action visualization
Each algorithm derives from The Trading Machine class and can be linked at runtime. Algorithms can be run in with real time data if an IB Client instantiation is provided or on historical data if a data source with recorded ticks (.tickdat) is provided. Algorithms are written and compiled in DLL. The main process in Qt Main will load the dll. Each algorithm exports a bool InitAlgorithm(std::string dataInput, IBInterfaceClient * ibInst). This function is loaded at runtime using GetProcAddress. This way, any algorithm can be designed and compiled as a DLL which can then be loaded and tested without the need to recompile the main framework.